Add a few tests
This patch adds Formol tests.

Each `.ft` file in the `tests` directory is made of two parts separated
by a `---` line: the top part is the input while the bottom part is the
expected output of formol.format().

Missing as of this date are ordered and definition list tests.

To run the tests:

    $ poetry install
    $ poetry run pytest

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>